Edit details from setting-page and keep homepage in view-only mode
