+++ site/public/js/server.js
@@ -1508,22 +1508,23 @@ function updateTheme(){
   detectColorScheme();
 }
 $(document).ready(function() {
-  if(localStorage.getItem("theme")){
-      if(localStorage.getItem("theme") === "dark"){
-        if(localStorage.getItem("black_mode") === "black"){
-          $("#theme_change_select").val("dark_black");
-        }else{
-          $("#theme_change_select").val("dark");
-        }
-      }else{
-        $("#theme_change_select").val("light");
+  let theme = "system";
+  let savedTheme = localStorage.getItem("theme");
+  let savedBlackMode = localStorage.getItem("black_mode");
+  if (savedTheme) {
+      if (savedTheme === "dark") {
+          theme = savedBlackMode === "black" ? "dark_black" : "dark";
+      } else {
+        theme = "light";
       }
-  }else{
-    if(localStorage.getItem("black_mode") === "black"){
-      $("#theme_change_select").val("system_black");
-    }else{
-      $("#theme_change_select").val("system");
-    }
+  } else {
+      theme = savedBlackMode === "black" ? "system_black" : "system";
+  }
+  // Now update the value of DOM element (if its profile-setting page drop-down value else its just text on home-page)
+  if ($("#theme_change_select").length) {
+    $("#theme_change_select").val(theme);
+  } else {
+    $("#theme_value").text(theme);
   }
 });
 
@@ -1589,7 +1590,7 @@ function resizeNoScrollTextareas() {
     // Make sure textareas resize correctly
     $('textarea.noscroll').each(function() {
         auto_grow(this);
-    })
+    });
 }
 
 $(document).ready(function() {
